+++ ssh/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fix creation of new tmux terminal windows
 - Add add-on icon
+- Update welcome logo
+- Fix SSH folder issue with authorized keys
 
 ## 8.1